refactor(cli): simplify workspace root guard
